# docs/api-reference/v1/openapi-v1.json
{
  "openapi": "3.0.3",
  "servers": [
    {
      "url": "https://api.cal.com/v1"
    }
  ],
...
  "info": {
    "title": "@calcom/api: Public API for Cal.com",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "apiKey"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/event-types/{id}": {
      "delete": {
        "operationId": "removeEventTypeById",
        "summary": "Remove an existing eventType",
        "parameters": [
          {
            "in": "query",
            "name": "apiKey",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Your API Key"
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID of the eventType to delete"
          }
        ],
        "tags": ["Event Types"],
        "externalDocs": {
          "url": "https://docs.cal.com/docs/core-features/event-types"
        },
        "responses": {
          "201": {
            "description": "OK, eventType removed successfully"
          },
          "400": {
            "description": "Bad request. EventType id is invalid."
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          }
        }
      },
      "get": {
        "operationId": "getEventTypeById",
        "summary": "Find a eventType",
        "parameters": [
          {
            "in": "query",
            "name": "apiKey",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Your API Key"
          },
          {
            "in": "path",
            "name": "id",
            "example": 4,
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID of the eventType to get"
          }
        ],
        "tags": ["Event Types"],
        "externalDocs": {
          "url": "https://docs.cal.com/docs/core-features/event-types"
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          },
          "404": {
            "description": "EventType was not found"
          }
        }
      },
      "patch": {
        "operationId": "editEventTypeById",
        "summary": "Edit an existing eventType",
        "parameters": [
          {
            "in": "query",
            "name": "apiKey",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Your API Key"
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID of the eventType to edit"
          }
        ],
        "requestBody": {
          "description": "Create a new event-type related to your user or team",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "length": {
                    "type": "integer",
                    "description": "Duration of the event type in minutes"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Metadata relating to event type. Pass {} if empty"
                  },
                  "title": {
                    "type": "string",
                    "description": "Title of the event type"
                  },
                  "slug": {
                    "type": "string",
                    "description": "Unique slug for the event type"
                  },
                  "scheduleId": {
                    "type": "number",
                    "description": "The ID of the schedule for this event type"
                  },
                  "hosts": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "number"
                        },
                        "isFixed": {
                          "type": "boolean",
                          "description": "Host MUST be available for any slot to be bookable."
                        }
                      }
                    }
                  },
                  "hidden": {
                    "type": "boolean",
                    "description": "If the event type should be hidden from your public booking page"
                  },
                  "position": {
                    "type": "integer",
                    "description": "The position of the event type on the public booking page"
                  },
                  "teamId": {
                    "type": "integer",
                    "description": "Team ID if the event type should belong to a team"
                  },
                  "periodType": {
                    "type": "string",
                    "enum": ["UNLIMITED", "ROLLING", "RANGE"],
                    "description": "To decide how far into the future an invitee can book an event with you"
                  },
                  "periodStartDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Start date of bookable period (Required if periodType is 'range')"
                  },
                  "periodEndDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "End date of bookable period (Required if periodType is 'range')"
                  },
                  "periodDays": {
                    "type": "integer",
                    "description": "Number of bookable days (Required if periodType is rolling)"
                  },
                  "periodCountCalendarDays": {
                    "type": "boolean",
                    "description": "If calendar days should be counted for period days"
                  },
                  "requiresConfirmation": {
                    "type": "boolean",
                    "description": "If the event type should require your confirmation before completing the booking"
                  },
                  "recurringEvent": {
                    "type": "object",
                    "description": "If the event should recur every week/month/year with the selected frequency",
                    "properties": {
                      "interval": {
                        "type": "integer"
                      },
                      "count": {
                        "type": "integer"
                      },
                      "freq": {
                        "type": "integer"
                      }
                    }
                  },
                  "disableGuests": {
                    "type": "boolean",
                    "description": "If the event type should disable adding guests to the booking"
                  },
                  "hideCalendarNotes": {
                    "type": "boolean",
                    "description": "If the calendar notes should be hidden from the booking"
                  },
                  "minimumBookingNotice": {
                    "type": "integer",
                    "description": "Minimum time in minutes before the event is bookable"
                  },
                  "beforeEventBuffer": {
                    "type": "integer",
                    "description": "Number of minutes of buffer time before a Cal Event"
                  },
                  "afterEventBuffer": {
                    "type": "integer",
                    "description": "Number of minutes of buffer time after a Cal Event"
                  },
                  "schedulingType": {
                    "type": "string",
                    "description": "The type of scheduling if a Team event. Required for team events only",
                    "enum": ["ROUND_ROBIN", "COLLECTIVE"]
                  },
                  "price": {
                    "type": "integer",
                    "description": "Price of the event type booking"
                  },
                  "currency": {
                    "type": "string",
                    "description": "Currency acronym. Eg- usd, eur, gbp, etc."
                  },
                  "slotInterval": {
                    "type": "integer",
                    "description": "The intervals of available bookable slots in minutes"
                  },
                  "successRedirectUrl": {
                    "type": "string",
                    "format": "url",
                    "description": "A valid URL where the booker will redirect to, once the booking is completed successfully"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the event type"
                  },
                  "seatsPerTimeSlot": {
                    "type": "integer",
                    "description": "The number of seats for each time slot"
                  },
                  "seatsShowAttendees": {
                    "type": "boolean",
                    "description": "Share Attendee information in seats"
                  },
                  "seatsShowAvailabilityCount": {
                    "type": "boolean",
                    "description": "Show the number of available seats"
                  },
                  "locations": {
                    "type": "array",
                    "description": "A list of all available locations for the event type",
                    "items": {
                      "type": "array",
                      "items": {
                        "oneOf": [
                          {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": ["integrations:daily"]
                              }
                            }
                          },
                          {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": ["attendeeInPerson"]
                              }
                            }
                          },
                          {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": ["inPerson"]
                              },
                              "address": {
                                "type": "string"
                              },
                              "displayLocationPublicly": {
                                "type": "boolean"
                              }
                            }
                          },
                          {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": ["link"]
                              },
                              "link": {
                                "type": "string"
                              },
                              "displayLocationPublicly": {
                                "type": "boolean"
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              },
              "example": {
                "event-type": {
                  "summary": "An example of event type PATCH request",
                  "value": {
                    "length": 60,
                    "requiresConfirmation": true
                  }
                }
              }
            }
          }
        },
        "tags": ["Event Types"],
        "externalDocs": {
          "url": "https://docs.cal.com/docs/core-features/event-types"
        },
        "responses": {
          "201": {
            "description": "OK, eventType edited successfully"
          },
          "400": {
            "description": "Bad request. EventType body is invalid."
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          }
        }
      }
    },
...
    "/selected-calendars": {
      "get": {
        "operationId": "listSelectedCalendars",
        "summary": "Find all selected calendars",
        "parameters": [
          {
            "in": "query",
            "name": "apiKey",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Your API Key"
          }
        ],
        "tags": ["Selected Calendars"],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          },
          "404": {
            "description": "No selected calendars were found"
          }
        }
      },
      "post": {
        "summary": "Creates a new selected calendar",
        "parameters": [
          {
            "in": "query",
            "name": "apiKey",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Your API Key"
          }
        ],
        "requestBody": {
          "description": "Create a new selected calendar",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["integration", "externalId"],
                "properties": {
                  "integration": {
                    "type": "string",
                    "description": "The integration name"
                  },
                  "externalId": {
                    "type": "string",
                    "description": "The external ID of the integration"
                  }
                }
              }
            }
          }
        },
        "tags": ["Selected Calendars"],
        "responses": {
          "201": {
            "description": "OK, selected calendar created"
          },
          "400": {
            "description": "Bad request. SelectedCalendar body is invalid."
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          }
        }
      }
    },
...
  }
}